fix(drafts): correct chat widget ownership
