--- apps/drafts/modules/social/models/chat/singlepage/list-default/figma.json
@@ -1,21 +1,0 @@
-{
-  "componentName": "social.chat",
-  "pageName": "social",
-  "nodeId": null,
-  "variantNodeId": null,
-  "variantName": "list-default",
-  "metadata": {
-    "sps.drafts.blockId": "social.chat.list-default",
-    "sps.drafts.layer": "singlepage",
-    "sps.drafts.syncKey": "social/models/chat/singlepage/list-default",
-    "sps.figma.component": "social.chat",
-    "sps.figma.variant": "list-default",
-    "sps.source.module": "social",
-    "sps.source.entityType": "model",
-    "sps.source.entity": "chat",
-    "sps.source.variant": "list-default",
-    "sps.contractVersion": "0.1.0",
-    "sps.code.component": "apps/drafts/modules/social/models/chat/singlepage/list-default/Component.tsx",
-    "sps.code.story": "apps/drafts/modules/social/models/chat/singlepage/list-default/Component.stories.tsx"
-  }
-}
