--- apps/drafts/modules/social/models/chat/singlepage/workspace-default/block.manifest.json
@@ -1,26 +1,0 @@
-{
-  "$schema": "../../../../../../block.schema.json",
-  "id": "social.chat.workspace-default",
-  "title": "Social Chat / Workspace / Default",
-  "description": "Singlepage presentation block for a chat workspace composed from social thread and message blocks.",
-  "layer": "singlepage",
-  "state": "draft",
-  "source": {
-    "module": "social",
-    "entityType": "model",
-    "entity": "chat",
-    "variant": "workspace-default"
-  },
-  "files": {
-    "component": "Component.tsx",
-    "story": "Component.stories.tsx"
-  },
-  "figma": {
-    "componentName": "social.chat",
-    "pageName": "social",
-    "nodeId": null,
-    "syncStatus": "not-created",
-    "metadataFile": "figma.json"
-  },
-  "tags": ["social", "chat", "thread", "message"]
-}
