--- apps/drafts/modules/social/models/chat/singlepage/workspace-default/figma.json
@@ -1,21 +1,0 @@
-{
-  "componentName": "social.chat",
-  "pageName": "social",
-  "nodeId": null,
-  "variantNodeId": null,
-  "variantName": "workspace-default",
-  "metadata": {
-    "sps.drafts.blockId": "social.chat.workspace-default",
-    "sps.drafts.layer": "singlepage",
-    "sps.drafts.syncKey": "social/models/chat/singlepage/workspace-default",
-    "sps.figma.component": "social.chat",
-    "sps.figma.variant": "workspace-default",
-    "sps.source.module": "social",
-    "sps.source.entityType": "model",
-    "sps.source.entity": "chat",
-    "sps.source.variant": "workspace-default",
-    "sps.contractVersion": "0.1.0",
-    "sps.code.component": "apps/drafts/modules/social/models/chat/singlepage/workspace-default/Component.tsx",
-    "sps.code.story": "apps/drafts/modules/social/models/chat/singlepage/workspace-default/Component.stories.tsx"
-  }
-}
