+++ apps/drafts/modules/social/models/widget/singlepage/chat-list-default/figma.json
@@ -1,0 +1,21 @@
+{
+  "componentName": "social.widget",
+  "pageName": "social",
+  "nodeId": null,
+  "variantNodeId": null,
+  "variantName": "chat-list-default",
+  "metadata": {
+    "sps.drafts.blockId": "social.widget.chat-list-default",
+    "sps.drafts.layer": "singlepage",
+    "sps.drafts.syncKey": "social/models/widget/singlepage/chat-list-default",
+    "sps.figma.component": "social.widget",
+    "sps.figma.variant": "chat-list-default",
+    "sps.source.module": "social",
+    "sps.source.entityType": "model",
+    "sps.source.entity": "widget",
+    "sps.source.variant": "chat-list-default",
+    "sps.contractVersion": "0.1.0",
+    "sps.code.component": "apps/drafts/modules/social/models/widget/singlepage/chat-list-default/Component.tsx",
+    "sps.code.story": "apps/drafts/modules/social/models/widget/singlepage/chat-list-default/Component.stories.tsx"
+  }
+}
