+++ apps/drafts/modules/social/models/widget/singlepage/chat-overview-default/figma.json
@@ -1,0 +1,21 @@
+{
+  "componentName": "social.widget",
+  "pageName": "social",
+  "nodeId": null,
+  "variantNodeId": null,
+  "variantName": "chat-overview-default",
+  "metadata": {
+    "sps.drafts.blockId": "social.widget.chat-overview-default",
+    "sps.drafts.layer": "singlepage",
+    "sps.drafts.syncKey": "social/models/widget/singlepage/chat-overview-default",
+    "sps.figma.component": "social.widget",
+    "sps.figma.variant": "chat-overview-default",
+    "sps.source.module": "social",
+    "sps.source.entityType": "model",
+    "sps.source.entity": "widget",
+    "sps.source.variant": "chat-overview-default",
+    "sps.contractVersion": "0.1.0",
+    "sps.code.component": "apps/drafts/modules/social/models/widget/singlepage/chat-overview-default/Component.tsx",
+    "sps.code.story": "apps/drafts/modules/social/models/widget/singlepage/chat-overview-default/Component.stories.tsx"
+  }
+}
