--- apps/drafts/modules/social/models/widget/singlepage/chat-workspace-default/Component.tsx
@@ -1,13 +1,0 @@
-import { SocialChatWorkspaceDefault } from "../../../chat/singlepage/workspace-default/Component";
-
-export function SocialWidgetChatWorkspaceDefault() {
-  return (
-    <section
-      data-ds-block="social.widget.chat-workspace-default"
-      data-ds-imports="social.chat.workspace-default"
-      data-ds-layer="singlepage"
-    >
-      <SocialChatWorkspaceDefault />
-    </section>
-  );
-}
