--- apps/drafts/modules/social/models/widget/singlepage/chat-workspace-default/figma.json
@@ -1,21 +1,0 @@
-{
-  "componentName": "social.widget",
-  "pageName": "social",
-  "nodeId": null,
-  "variantNodeId": null,
-  "variantName": "chat-workspace-default",
-  "metadata": {
-    "sps.drafts.blockId": "social.widget.chat-workspace-default",
-    "sps.drafts.layer": "singlepage",
-    "sps.drafts.syncKey": "social/models/widget/singlepage/chat-workspace-default",
-    "sps.figma.component": "social.widget",
-    "sps.figma.variant": "chat-workspace-default",
-    "sps.source.module": "social",
-    "sps.source.entityType": "model",
-    "sps.source.entity": "widget",
-    "sps.source.variant": "chat-workspace-default",
-    "sps.contractVersion": "0.1.0",
-    "sps.code.component": "apps/drafts/modules/social/models/widget/singlepage/chat-workspace-default/Component.tsx",
-    "sps.code.story": "apps/drafts/modules/social/models/widget/singlepage/chat-workspace-default/Component.stories.tsx"
-  }
-}
